refactor: move load more functionality to ItemList component
- ItemList now owns its data loading logic and state management
- Simplified FolderPage by removing items, hasMore, isLoading state
- ItemList reads order from URL via loadMoreItems server action
- Better separation of concerns with self-contained components
- Improved component reusability and maintainability
