+++ components/ItemList/ItemList.tsx
@@ -1,30 +1,58 @@
 "use client";
 
-import { useEffect, useState } from "react";
+import { useEffect, useState, useCallback } from "react";
 import { Gallery } from "react-photoswipe-gallery";
 import { useInView } from "react-intersection-observer";
 import type { Item, Layout } from "@/types/models";
 import { ItemItem } from "./ItemItem";
+import { loadMoreItems, type ItemsPage } from "@/actions/items";
 import styles from "./ItemList.module.css";
 import "photoswipe/dist/photoswipe.css";
 
 interface ItemListProps {
-  items: Item[];
+  folderId: string;
+  initialItemsPage: ItemsPage;
   libraryPath: string;
   layout: Layout;
-  hasMore?: boolean;
-  isLoading?: boolean;
-  onLoadMore?: () => void;
 }
 
 export function ItemList({
-  items,
+  folderId,
+  initialItemsPage,
   libraryPath,
   layout,
-  hasMore = false,
-  isLoading = false,
-  onLoadMore,
 }: ItemListProps) {
+  const [items, setItems] = useState<Item[]>(initialItemsPage.items);
+  const [hasMore, setHasMore] = useState(initialItemsPage.hasMore);
+  const [isLoading, setIsLoading] = useState(false);
+
+  // Reset items when initial data changes
+  useEffect(() => {
+    setItems(initialItemsPage.items);
+    setHasMore(initialItemsPage.hasMore);
+  }, [initialItemsPage.items, initialItemsPage.hasMore]);
+
+  // Load more items callback
+  const handleLoadMore = useCallback(async () => {
+    if (isLoading || !hasMore) return;
+
+    setIsLoading(true);
+    try {
+      const itemsPage = await loadMoreItems({
+        folderId,
+        offset: items.length,
+        limit: 100,
+      });
+
+      setItems((prev) => [...prev, ...itemsPage.items]);
+      setHasMore(itemsPage.hasMore);
+    } catch (error) {
+      console.error("Failed to load more items:", error);
+    } finally {
+      setIsLoading(false);
+    }
+  }, [folderId, items.length, hasMore, isLoading]);
+
   // Set rootMargin to 50% of viewport height for eager loading
   const rootMargin =
     typeof window !== "undefined"
@@ -37,11 +65,12 @@ export function ItemList({
 
   const [lastInView, setLastInView] = useState(false);
   useEffect(() => {
-    if (!lastInView && inView && hasMore && !isLoading && onLoadMore) {
-      onLoadMore();
+    if (!lastInView && inView && hasMore && !isLoading) {
+      handleLoadMore();
     }
     setLastInView(inView);
-  }, [lastInView, inView, hasMore, isLoading, onLoadMore]);
+  }, [lastInView, inView, hasMore, isLoading, handleLoadMore]);
+
   return (
     <Gallery
       withCaption
